refactor(lab): extract offload workspace planning
